requirements.txt
streamlit
tensorflow
keras
numpy
gtts
Pillow
